backend/requirements.txt
python-chess>=1.999
uvicorn
python-dotenv
